gemini/sample-apps/quickbot/multi-agent-travel-concierge-with-adk/backend/requirements.txt
poetry
gunicorn
uvicorn[standard]
websocket
fastapi
fastapi-utilities
cloudpickle
h5py
grpcio
urllib3
typing-inspect
typing_extensions
pydantic
requests
langchain
langchain-community
langchain-core
langchain-google-vertexai
google-adk==0.5.0
google-cloud-aiplatform[adk, agent_engines]==1.88.0
google-cloud-bigquery
google-cloud-tasks
google-cloud-logging
google-cloud-speech
python-multipart